requirements.txt
annotated-types==0.7.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pydantic==2.4.2
pydantic_core==2.10.1
SQLAlchemy==2.0.20
typing_extensions==4.12.2
Werkzeug==3.1.3
